nextgen-route-migration/requirements.txt
pathlib==1.0.1 ; python_version == '3.8'
PyYAML==6.0 ; python_version == '3.8'